Cleaned up the G2-Control schematic.
Removed Netclass names from their flags and greyed the flags so thet they
interfere less with the actual schematic.

Moved things around to avoid collisions.
Moved D6 into the on-board space and shrunk the cutout for R12 and Q2
Moved BR1 to get less parallel wirings for the same signal.

Changed the SCR symbol in the library to make it more clear.
Shortened the gate pin of Q2 to free some vertical space.

Changed the values of some components that need to be configured to have an asterisk instead of a default value.