Replaced Q4 transistor schematic symbol BC557 with 2N3904 in G2-Control schematic
Updated the G2-Control schematic by replacing the schematic symbol BC557 for transistor Q4 with the symbol for 2N3904 which has the same pin numbering as the actual transistor used. This change fixes
issue #11 and makes the silkscreen print on the PCB show the transistor the way it has to be rotated in order to function properly.

Other changes is that the references and values of the components have been moved and the valye is set in a smaller font to increase readability. The global labels has also been replaced by hierarchical labels.