Updated relay components and adjusted wiring
The K1 relay has been updated from model AZ822-2C-24DSE to RTE24024 for better current handling in case of arc-over. Additionally, wiring for these components has been adjusted to fit the new model.
Moved R13 to a better location and increased its footprint.

Moved the grounding hole to a place near both the VDR2 and the connection to K1 in order to get as low as possible impedance to the cathode connection for these two components.